Move filters namespace one directory up
