+++ typings/common.ts
@@ -1,0 +1,4 @@
+export type Genre = {
+    id: number;
+    name: string;
+  };
