+++ typings/filters.ts
@@ -1,9 +1,9 @@
-import type {movie} from './movie';
+import { Genre } from './common';
 
 export namespace filters {
   export type SortOption = { value: string; label: string };
 
-  export type GenreList = (movie.Genre & {
+  export type GenreList = (Genre & {
     selected?: boolean;
   })[];
 }
